Move utility modules to lib folder
