+++ lib/jwt.js
